moves tabset + tab components; moves animations.ts
